#7: Improve inclusion and exclusion of repos
This improves the inclusion and exclusion of repos.

The dependency for the command line args parser has been changed from
commander to yargs to more easily support multi-argument options.

The command line API has been changed for the tool. Any usage of the old
versions will fail as the --org argument has been changed to a
postional. This will, as a byproduct, make sure that any current usage
of the script is updated to use the correct arguments.